+++ .env.testing
@@ -83,7 +83,7 @@ POSTGRES_DB=trunk_transcribe
 #
 
 # This should be a publicly-accessible URL to the API server, from whever the worker is being run
-API_BASE_URL=http://api:8000
+API_BASE_URL=http://api:8000/api/v1
 API_KEY=testing
 
 #
