+++ frontend/.env.example
@@ -1,7 +1,7 @@
 # This should be a publicly-accessible URL to the API server, from whever the worker is being run
-VITE_WEBSOCKET_URL=ws://localhost:8000/ws
+VITE_WEBSOCKET_URL=ws://localhost:8000/api/v1/ws
 VITE_API_KEY=testing
-VITE_API_BASE_URL=http://localhost:8000
+VITE_API_BASE_URL=http://localhost:8000/api/v1
 CHAT_AGENT_URL=http://localhost:7932
 
 VITE_MEILI_MASTER_KEY=testing
